requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
minio==7.1.3
python-multipart==0.0.6
pydantic==1.10.11
SQLAlchemy==1.4.52
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
pydantic[email]
